TIL/03.23/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.94
ultralytics-thop==2.0.14
urllib3==2.3.0
